backend_nexo/pyproject.toml
[tool.poetry]
name = "backend-nexo-"
version = "0.1.0"
description = ""
authors = ["maximilianofni <maximilianofni@gmail.com>"]
readme = "README.md"
packages = [{include = "backend_nexo"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.5"
psycopg2 = "^2.9.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
